backend/backend/requirements.txt
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
requests==2.32.3

google~=3.0.0